chore: move atauth code in atcommons into atauth
