+++ packages/atauth/src/atactivate.c
@@ -10,7 +10,7 @@
 #include <atchops/aes.h>
 #include <atchops/aes_ctr.h>
 #include <atchops/iv.h>
-#include <atcommons/enroll_status.h>
+#include <atauth/enroll_status.h>
 #include <atlogger/atlogger.h>
 
 #include <atclient/string_utils.h>
