+++ packages/atauth/src/enroll_command_builder.c
@@ -1,6 +1,6 @@
-#include "atcommons/enroll_command_builder.h"
-#include "atcommons/enroll_operation.h"
-#include "atcommons/enroll_params.h"
+#include "atauth/enroll_command_builder.h"
+#include "atauth/enroll_operation.h"
+#include "atauth/enroll_params.h"
 #include "atlogger/atlogger.h"
 
 #include <stddef.h>
