+++ packages/atauth/tests/test_cjson.c
@@ -1,4 +1,4 @@
-#include <atcommons/json.h>
+#include <atclient/json.h>
 #include <stddef.h> // IWYU pragma: keep
 #include <stdio.h>
